refactor: move server-side image processing to API route
- Separated client and server-side image processing code
- Created new image API route for server operations
- Fixed module resolution errors in build
